+++ packages/remote-ontology/README.md
@@ -4,12 +4,11 @@ HTTP remoting for LiveOntology with policy-aware describe/load/apply and structu
 
 ## Describe projection
 
-`projectRemoteOntologyIR` supports:
-
-- `projectionMode: "legacy"` — historical behavior (full object/link schema; action fixed-parameter projection)
-- `projectionMode: "authorized"` — project object types/properties, prune broken links, hide actions/queries that reference hidden types, and drop unreachable named types
-
-When `allowedObjectTypeProperties` or `baseObjectTypeQueries` is configured, describe defaults to `authorized`.
+Describe retains the full object and link schema when no object-visibility
+policy is configured. When `allowedObjectTypeProperties` or
+`baseObjectTypeQueries` is configured, it automatically projects authorized
+object types and properties, prunes broken links, hides actions and queries
+that reference hidden types, and drops unreachable named types.
 
 ## Errors
